Hypervisor: perform runs before fuzzing, input dir '-'
Now setting corpus to mode Normal requires the total coverage of the
seed corpus. This way, corpus recorded coverage now actually corresponds
to the coverage of its inputs. Before, we always added the first
mutations to the corpus because the coverage was empty, despite we may
already had a seed input for that new coverage.
Now using '-' as input directory will set it as the output corpus
directory, so the output corpus of a previous fuzz session can be
reused as seed corpus of a new session.