Run timeouts
Timeouts are handled by the kernel. It implements a timer using the periodic LAPIC interrupt. If it exceeds a certain value, which can be changed from the hypervisor, the run will end with RunEndReason::Timeout
Kernel now uses hc_end_run every time the run ends, giving the reason as argument, instead of having different hypercalls for each reason.