Kernel zig: trying new allocator and scheduling/multitasking
The way multitasking works right now is having a single kernel stack.
Task switching is done replacing the registers in the interrupt/syscall
frame with the ones of the process we are switching to, and loading its
page table if needed.
As there is a single kernel stack for all processes, the kernel is now
non-preemptable (interrupts are disabled while in kernel space). I think
this is not the best way of doing it, but we'll see :)