Kernel: (failed) attempt to implement multitasking
While writing this I was very uncomfortable with the amount of code that didn't check errors and would crash on OOM. I added an Optional type inspired by SerenityOS while trying to solve this, but it was just a small patch. This felt like a great opportunity to rewrite the kernel in Zig, and so I did. This kernel is now deprecated.