Stacktraces and switch build system to Zig
Stacktraces are one honking great idea -- let's do more of those! Now every
unique crash prints the stacktrace that caused the fault. If the binary has
debugging information, it will print symbol and source information associated to
each address. As this is implemented in the hypervisor, this works for the
fuzzed target but also for the kernel, so we don't need to print stacktraces
when panicking in the kernel: we just need to submit a crash to the hypervisor.

Builds seem to be a bit slower with Zig than with gcc&cmake. However, I love
having all build instructions in a single script, and having Zig as the only
dependency for building (besides hypervisor libraries). Also, passing options to
the kernel is now easier. On the other side, it seems that the userspace tests
binary is not correctly built with Zig (or maybe some semantics changed with new
libc?). This may be related: https://github.com/ziglang/zig/issues/10386