Stacktraces for libraries, hook breakpoints
ElfParser:
	- Remove m_base, keep things simpler with m_load_addr
	- Add a static method for getting a stacktrace using debug info of
	several elfs

Added hook breakpoints

Keep track of dynamic libraries. Now each time mmap is called to map a file, the
kernel checks if it is being called from the loader. In that case, it notifies
the hypervisor so it can parse the library and get access to its symbols and
debug information. This means we can now provide a stacktrace with symbol and
source information even for libraries and dynamically linked binaries.