Hypervisor: fix Intel PT coverage
Previously, KVM_VMX_PT_CHECK_TOPA_OVERFLOW ioctl was used at the end of
each run. I thought this reset the TOPA and returned the trace size even
if there was not overflow, but it didn't. The other option for getting
the trace size and resetting was calling KVM_VMX_PT_DISABLE, as QEMU-Nyx
does, but then we would need to enter the VM and only then enable VMX PT
again with a hypercall. This isn't possible in our case, since the VM is
restored to a state in userland. So the solution is to fork KVM-Nyx and
implement a simple ioctl KVM_VMX_PT_RESET that just resets the TOPA and
returns the trace size.