Fix instruction count
Previously, at the end of each run the kernel read the MSR and told the
hypervisor the number of instructions executed. However, when the run
ends at a breakpoint, the kernel doens't execute and it isn't able to do
that. Now, we read the MSR from the hypervisor after each run.