Kernel: refcounting and starting to fix leaks
Added RefCountType to utils.RefCounter
Added reference counting to AddressSpace
Rewritten PageTableEntry so it uses a packed struct with each field instead of working manually with the underlying usize
Added reference counting to physical memory. Since we now have multiple processes, we can not just free the frame when unmapping some page, because it may be mapped multiple times.